requirements.txt
fastapi
uvicorn[standard]
python-multipart
passlib[bcrypt]
python-jose
meshpy